refactor: change Infrastructure to Data
