+++ Backend/CampusTradeSystem/CampusTrade.Tests/IntegrationTests/ApiEndToEndTests.cs
@@ -5,7 +5,6 @@ using System.Text;
 using System.Text.Json;
 using CampusTrade.API;
 using CampusTrade.API.Data;
-using CampusTrade.API.Infrastructure;
 using CampusTrade.API.Models.DTOs.Auth;
 using CampusTrade.API.Models.DTOs.Common;
 using CampusTrade.API.Services.BackgroundServices;
