+++ Backend/CampusTradeSystem/CampusTrade.Tests/IntegrationTests/AuthIntegrationTests.cs
@@ -4,7 +4,6 @@ using System.Text;
 using System.Text.Json;
 using CampusTrade.API;
 using CampusTrade.API.Data;
-using CampusTrade.API.Infrastructure;
 using CampusTrade.API.Models.DTOs.Auth;
 using CampusTrade.API.Models.DTOs.Common;
 using CampusTrade.API.Models.Entities;
